+++ extension/src/contentFunctions/video/css.ts
@@ -1,0 +1,58 @@
+import { CSS_IDS } from "../../enums";
+
+/**
+ * AddVideoCSS
+ * 1. Create the style element
+ * 1.1 Set the id
+ * 1.2 Add a glow to the video box
+ * 2. Add the style to the page
+ */
+export const addVideoCSS = (): Promise<void> => {
+  // 1.
+  // Create the style element
+  // eslint-disable-next-line no-undef
+  const style = document.createElement("style");
+  // 1.1
+  // Set the id
+  style.id = CSS_IDS.VIDEO;
+  // 1.2
+  // Add a glow to the video box
+  style.innerHTML = `#player {
+  transition-delay: 0.5s;
+  transition: box-shadow 0.6s cubic-bezier(0.165, 0.84, 0.44, 1);
+  box-shadow: -10px 0 40px rgb(62, 187, 243), 10px 0 40px rgb(88, 80, 209);
+}
+
+#movie_player > div.html5-video-container > video {
+  transition-delay: 0.5s;
+  transition: box-shadow 0.6s cubic-bezier(0.165, 0.84, 0.44, 1);
+  box-shadow: -10px 0 10vw rgb(62, 187, 243), 10px 0 10vw rgb(88, 80, 209);
+  clip-path: inset(0px -100vw 0px -100vw);
+}`;
+
+  // 2.
+  // Add the style to the page
+  // eslint-disable-next-line no-undef
+  document.head.appendChild(style);
+
+  return Promise.resolve();
+};
+
+/**
+ * RemoveVideoCSS
+ * 1. Get the style element
+ * 2. Remove the style element
+ */
+export const removeVideoCSS = (): Promise<void> => {
+  // 1.
+  // Get the style element
+  // eslint-disable-next-line no-undef
+  const style = document.getElementById(CSS_IDS.VIDEO);
+  if (style) {
+    // 2.
+    // Remove the style element
+    style.remove();
+  }
+
+  return Promise.resolve();
+};
